2d grid map, wsad and mouse imput added
