+++ include/cube3d.h
@@ -40,7 +40,6 @@ typedef struct s_map
 	char	**map_ids;
 	char	***config;
 	char	*flat_map;
-	int		n_rows;
 	char	**map_array;
     int     height;
     int     width;
@@ -94,17 +93,17 @@ void	clear_image_to_colour(t_data *data, int colour);
 void	free_2d_char_arr(char **arr);
 
 // load_map
-//	static int		init_ids(t_data *data);
-//	static int	parse_line(t_data *data, char *line)
 int		load_map_data(t_data *data, char *f_name);
 int		validate_map(t_data *data);
+//	static int		init_ids(t_data *data);
+//	static int	parse_line(t_data *data, char *line)
 
 // load_map_utils
-// static int		parse_map(t_data *data, char *line);
 int		free_map_data(t_data *data);
 bool	member_of_set(char c, char *set);
 int		free_temp_return(char **temp, int r);
 int		array_len(char **arr);
+// static int		parse_map(t_data *data, char *line);
 
 // graphics_image
 void	init_img(t_data *data);
@@ -116,6 +115,13 @@ void	clear_image_to_colour(t_data *data, int colour);
 // graphics_lines
 t_line	*init_line(int x1, int y1, int x2, int y2);
 void	compute_line_points(t_data *data, t_line *line);
+
+// graphics_grid
+int		draw_grid(t_data *data);
+// static void	fill_square(t_data *data, int x, int y, int len_side)
+
+// keyboard_input
+int	key_down(int keycode, t_data *data);
 
 // get_next_line
 char	*get_next_line(int fd);
