+++ src/debug.c
@@ -1,4 +1,4 @@
-#include "window.h"
+#include "cube3d.h"
 
 void print_string_array(char **array, int n) {
     for (int i = 0; i < n; i++) {
