+++ src/error_msg.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "window.h"
+#include "cube3d.h"
 
 int error(char *message)
 {
