+++ src/get_next_line.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "window.h"
+#include "cube3d.h"
 
 char	*get_next_line(int fd)
 {
