+++ src/get_next_line_utils.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "window.h"
+#include "cube3d.h"
 
 char	*set(char **ln, int fd)
 {
